number_plate_detection/requirements.txt
opencv-python
streamlit
numpy
easyocr
pillow
